@@ -1,9 +1,9 @@
 package com.yzx.delegate.holder;
 
 import android.graphics.Bitmap;
-import android.support.annotation.DrawableRes;
-import android.support.annotation.IdRes;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.DrawableRes;
+import androidx.annotation.IdRes;
+import androidx.recyclerview.widget.RecyclerView;
 import android.text.Spanned;
 import android.util.SparseArray;
 import android.view.View;
