@@ -1,7 +1,7 @@
 package com.yzx.delegate.items;
 
 import android.content.Context;
-import android.support.annotation.LayoutRes;
+import androidx.annotation.LayoutRes;
 import android.util.SparseArray;
 
 import com.yzx.delegate.DelegateManager;
